# Deflate speed up