--- a/src/lz77.ts
+++ b/src/lz77.ts
@@ -46,10 +46,11 @@ export function generateLZ77CodeValues(input: Uint8Array, indexMap: {[key: numbe
       nowIndex++;
       continue;
     }
-    indexes.forEach((hidIndex) => {
-      if (slideIndexBase <= hidIndex && hidIndex < nowIndex ) {
+
+    for (let i = 0, iMax = indexes.length; i < iMax; i++) {
+      if (slideIndexBase <= indexes[i] && indexes[i] < nowIndex ) {
         repeatLength = 0;
-        while (input[hidIndex + repeatLength] === input[nowIndex + repeatLength]) {
+        while (input[indexes[i] + repeatLength] === input[nowIndex + repeatLength]) {
           repeatLength++;
           if (257 < repeatLength) {
             break;
@@ -57,10 +58,10 @@ export function generateLZ77CodeValues(input: Uint8Array, indexMap: {[key: numbe
         }
         if (repeatLengthMax < repeatLength) {
           repeatLengthMax = repeatLength;
-          repeatLengthMaxIndex = hidIndex;
+          repeatLengthMaxIndex = indexes[i];
         }
       }
-    });
+    }
     if (repeatLengthMax >= 3) {
       distance = nowIndex - repeatLengthMaxIndex;
       for (let i = 0; i < LENGTH_EXTRA_BIT_BASE.length; i++) {