Improve deflate speed.
Speed is measured by binary data test.

### Changed LZ77 algorithm to Index search

|No.|before|after|
|:---:|---:|---:|
|1|35379 ms|18914 ms|
|2|35141 ms|19090 ms|
|3|35515 ms|18949 ms|
|4|35428 ms|19007 ms|
|5|35485 ms|19019 ms|

### 

|No.|after|
|:---:|---:|
|1|ms|
|2|ms|
|3|ms|
|4|ms|
|5|ms|